backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pandas
httpx
asyncpg
sqlalchemy
supabase